split Attention and Transformer
